backend/pyproject.toml
[project]
name = "cloudless-backend"
version = "1.0.0"
description = "Secure self-hosted file transfer backend with E2E encryption"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.3",
    "argon2-cffi>=25.1.0",
    "fastapi>=0.128.4",
    "pydantic-settings>=2.12.0",
    "pynacl>=1.6.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn[standard]>=0.40.0",
]

[project.scripts]
cloudless = "uvicorn:main"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]
